Factors out joystick logic.
