+++ Machines/Apple/AppleII/Joystick.cpp
@@ -1,0 +1,9 @@
+//
+//  Joystick.cpp
+//  Clock Signal
+//
+//  Created by Thomas Harte on 16/02/2021.
+//  Copyright © 2021 Thomas Harte. All rights reserved.
+//
+
+#include "Joystick.hpp"
